Add running/walking state lock feature. Add dashing mode to the player. Update config Ui

+++ app/src/main/java/com/example/myapplication/Model/loopVideo/GameAnimation.java
@@ -5,7 +5,7 @@ import android.graphics.PointF;
 public class GameAnimation extends VideoFrame {
 
 
-    public GameAnimation(int locX, int locY, int width, int height, GameVideos anim) {
+    public GameAnimation(float locX, float locY, int width, int height, GameVideos anim) {
         super(new PointF(locX, locY), width, height, anim);
 
     }
